requirements.txt
aiohttp>=3.9.0
asyncpg>=0.29.0
python-telegram-bot>=20.7
pyyaml>=6.0
pytz>=2024.1
numpy>=1.26.0
python-dateutil>=2.8.0
uvloop>=0.19.0; sys_platform != "win32"
